Allow a NULL header in PEM_write{,_bio}()
When PEM_write{,_bio}() were documented by Rich Salz and Richard Levitte,
it was incorrectly stated that the header argument is allowed to be NULL.
This was never true.  Instead of fixing the documentation, it was decided
that the API needs a fix, so pull in a variant of OpenSSL 3b9082c8.

ok jsing

(cherry picked from commit 4526d7fe8f394574c67222cfc9ff3026791d0734)