blist: fix a possible blist corruption with blist_alloc() due to unsigned swblk_t on OpenBSD.
reorder if condition in blst_meta_alloc(), in order to check if the node is
'Terminator' node first (and leave the loop).

DragonFlyBSD is unaffected by it as swblk_t is signed (and the first condition
isn't taken).

add a regress test for it.

while here, more the KASSERT() to KDASSERT(). it is useful but only with DEBUG.

ok miod@ todd@

(cherry picked from commit 31c3ffda2848f054f38e2e590ed853f201b7a5e0)